Resync the SVN cache just before using it when checking if a folder exists, rather than just after creating the folder.
Still a bit of a race condition I suppose - better to just try to create
the folders and fail if they already exist.
Even in doing that case, I've seen the writer sometimes fail if the
repository is newer than it's idea of "youngest". I don't fully know how
to do this.
Maybe fixes #3301

git-svn-id: file:///mnt/ramdisk/d4-svn/trunk@10382 484df14f-75d2-4b3f-8ba0-490e05069a01